Feat: 가장 가까운 적 탐색 후 타겟 설정 Feat: 사거리 확인 후 공격

+++ Assets/Scripts/Piece/PieceData.cs
@@ -17,7 +17,7 @@ public class PieceData : ScriptableObject
     [SerializeField] float attackSpeed;
     [SerializeField] float criticalChance;
     [SerializeField] float criticalDamage;
-    [SerializeField] float attackRange;
+    [SerializeField] int attackRange;
 
     public enum Mythology { NONE = -1, A, B, C, D, E, MAX }
     public enum Species { NONE = -1, HAMSTER, CAT, DOG, FROG, RABBIT, MAX }
